Move user id retrieval logic to processor
