+++ djangosaml2idp/processors.py
@@ -1,3 +1,6 @@
+from django.conf import settings
+
+
 class BaseProcessor:
     """ Processor class is used to determine if a user has access to a client service of this IDP
         and to construct the identity dictionary which is sent to the SP
@@ -19,6 +22,11 @@ class BaseProcessor:
         """ Check if this user should use a second authentication system
         """
         return False
+
+    def get_user_id(self, user):
+        user_field = getattr(settings, 'SAML_IDP_DJANGO_USERNAME_FIELD', None) or \
+                     getattr(user, 'USERNAME_FIELD', 'username')
+        return str(getattr(user, user_field))
 
     def create_identity(self, user, sp_mapping, **extra_config):
         """ Generate an identity dictionary of the user based on the given mapping of desired user attributes by the SP
