+++ tests/test_views.py
@@ -18,24 +18,6 @@ class TestIdPHandlerViewMixin:
     def test_get_identity_provides_extra_config(self):
         obj = IdPHandlerViewMixin()
 
-    def test_extract_user_id_configure_by_user_class(self):
-
-        user = User()
-        user.USERNAME_FIELD = 'email'
-        user.email = 'test_email'
-
-        assert IdPHandlerViewMixin().extract_user_id(user) == 'test_email'
-
-    def test_extract_user_id_configure_by_settings(self, settings):
-        """Should use `settings.SAML_IDP_DJANGO_USERNAME_FIELD` to determine the user id field"""
-
-        settings.SAML_IDP_DJANGO_USERNAME_FIELD = 'first_name'
-
-        user = User()
-        user.first_name = 'test_first_name'
-
-        assert IdPHandlerViewMixin().extract_user_id(user) == 'test_first_name'
-
     def test_get_processor_errors_if_processor_cannot_be_loaded(self):
         sp_config = {
             'processor': 'this.does.not.exist'
